(PC-24550)[PRO] feat: focus first input on signin api error
